Move Sarama YAML related functions into common
