+++ pkg/channel/distributed/common/kafka/sarama/sarama.go
@@ -24,7 +24,6 @@ import (
 	"io/ioutil"
 	"log"
 	"os"
-	"regexp"
 
 	"github.com/Shopify/sarama"
 	"github.com/ghodss/yaml"
@@ -35,12 +34,10 @@ import (
 	commonconfig "knative.dev/eventing-kafka/pkg/channel/distributed/common/config"
 	"knative.dev/eventing-kafka/pkg/channel/distributed/common/kafka/constants"
 	"knative.dev/eventing-kafka/pkg/channel/distributed/common/testing"
+	"knative.dev/eventing-kafka/pkg/common/client"
 	kubeclient "knative.dev/pkg/client/injection/kube/client"
 	"knative.dev/pkg/system"
 )
-
-// Regular Expression To Find All Certificates In Net.TLS.Config.RootPEMs Field
-var regexRootPEMs = regexp.MustCompile(`(?s)\s*RootPEMs:.*-----END CERTIFICATE-----`)
 
 // Utility Function For Enabling Sarama Logging (Debugging)
 func EnableSaramaLogging(enable bool) {
@@ -66,139 +63,6 @@ func UpdateSaramaConfig(config *sarama.Config, clientId string, username string,
 
 	// We Always Want Success Messages From Producer
 	config.Producer.Return.Successes = true
-}
-
-//
-// Extract (Parse & Remove) Top Level Kafka Version From Specified Sarama Confirm YAML String
-//
-// The Sarama.Config struct contains a top-level 'Version' field of type sarama.KafkaVersion.
-// This type contains a package scoped [4]uint field called 'version' which we cannot unmarshal
-// the yaml into.  Therefore, we instead support the user providing a 'Version' string of the
-// format "major.minor.patch" (e.g. "2.3.0") which we will "extract" out of the YAML string
-// and return as an actual sarama.KafkaVersion.  In the case where the user has NOT specified
-// a Version string we will return a default version.
-//
-func extractKafkaVersion(saramaConfigYamlString string) (string, sarama.KafkaVersion, error) {
-
-	// Define Inline Struct To Marshall The Top Level Sarama Config "Kafka" Version Into
-	type saramaConfigShell struct {
-		Version string
-	}
-
-	// Unmarshal The Sarama Config Into The Shell
-	shell := &saramaConfigShell{}
-	err := yaml.Unmarshal([]byte(saramaConfigYamlString), shell)
-	if err != nil {
-		return saramaConfigYamlString, constants.ConfigKafkaVersionDefault, err
-	}
-
-	// Return Default Kafka Version If Not Specified
-	if len(shell.Version) <= 0 {
-		return saramaConfigYamlString, constants.ConfigKafkaVersionDefault, nil
-	}
-
-	// Attempt To Parse The Version String Into A Sarama.KafkaVersion
-	kafkaVersion, err := sarama.ParseKafkaVersion(shell.Version)
-	if err != nil {
-		return saramaConfigYamlString, constants.ConfigKafkaVersionDefault, err
-	}
-
-	// Remove The Version From The Sarama Config YAML String
-	regex, err := regexp.Compile("\\s*Version:\\s*" + shell.Version + "\\s*")
-	if err != nil {
-		return saramaConfigYamlString, constants.ConfigKafkaVersionDefault, err
-	} else {
-		updatedSaramaConfigYamlBytes := regex.ReplaceAll([]byte(saramaConfigYamlString), []byte{})
-		return string(updatedSaramaConfigYamlBytes), kafkaVersion, nil
-	}
-}
-
-/* Extract (Parse & Remove) TLS.Config Level RootPEMs From Specified Sarama Confirm YAML String
-
-The Sarama.Config struct contains Net.TLS.Config which is a *tls.Config which cannot be parsed.
-due to it being from another package and containing lots of func()s.  We do need the ability
-however to provide Root Certificates in order to avoid having to disable verification via the
-InsecureSkipVerify field.  Therefore, we support a custom 'RootPEMs' field which is an array
-of strings containing the PEM file content.  This function will "extract" that content out
-of the YAML string and return as a populated *x509.CertPool which can be assigned to the
-Sarama.Config.Net.TLS.Config.RootCAs field.  In the case where the user has NOT specified
-any PEM files we will return nil.
-
-In order for the PEM file content to pass cleanly from the YAML string, which is itself
-already a YAML string inside the K8S ConfigMap, it must be formatted and spaced correctly.
-The following shows an example of the expected usage...
-
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: config-kafka
-  namespace: knative-eventing
-data:
-  sarama: |
-    Admin:
-      Timeout: 10000000000
-    Net:
-      KeepAlive: 30000000000
-      TLS:
-        Enable: true
-        Config:
-          RootCaPems:
-          - |-
-            -----BEGIN CERTIFICATE-----
-            MIIGBDCCA+ygAwIBAgIJAKi1aEV58cQ1MA0GCSqGSIb3DQEBCwUAMIGOMQswCQYD
-            ...
-            2wk9rLRZaQnhspt6MhlmU0qkaEZpYND3emR2XZ07m51jXqDUgTjXYCSggImUsARs
-            NAehp9bMeco=
-            -----END CERTIFICATE-----
-      SASL:
-        Enable: true
-
-...where you should make sure to use the YAML string syntax of "|-" in order to
-prevent trailing linefeed. The indentation of the PEM content is also important
-and must be aligned as shown.
-*/
-func extractRootCerts(saramaConfigYamlString string) (string, *x509.CertPool, error) {
-
-	// Define Inline Struct To Marshall The TLS Config 'RootPEMs' Into
-	type tlsConfigShell struct {
-		Net struct {
-			TLS struct {
-				Config struct {
-					RootPEMs []string
-				}
-			}
-		}
-	}
-
-	// Unmarshal The TLS Config Into The Shell
-	shell := &tlsConfigShell{}
-	err := yaml.Unmarshal([]byte(saramaConfigYamlString), shell)
-	if err != nil {
-		return saramaConfigYamlString, nil, err
-	}
-
-	// Convenience Variable For The RootPEMs
-	rootPEMs := shell.Net.TLS.Config.RootPEMs
-
-	// Exit Early If No RootCert PEMs
-	if rootPEMs == nil || len(rootPEMs) < 1 {
-		return saramaConfigYamlString, nil, nil
-	}
-
-	// Create A New CertPool To Contain The Root PEMs
-	certPool := x509.NewCertPool()
-
-	// Populate The CertPool With The PEMs
-	for _, rootPEM := range rootPEMs {
-		ok := certPool.AppendCertsFromPEM([]byte(rootPEM))
-		if !ok {
-			return saramaConfigYamlString, nil, fmt.Errorf("failed to parse root certificate PEM: %s", rootPEM)
-		}
-	}
-
-	// Remove The RootPEMs From The Sarama YAML String (Multi-Line / Greedy To Collect All PEMs)
-	updatedSaramaConfigYamlBytes := regexRootPEMs.ReplaceAll([]byte(saramaConfigYamlString), []byte{})
-	return string(updatedSaramaConfigYamlBytes), certPool, nil
 }
 
 // ConfigEqual is a convenience function to determine if two given sarama.Config structs are identical aside
@@ -260,13 +124,13 @@ func MergeSaramaSettings(config *sarama.Config, configMap *corev1.ConfigMap) (*s
 	saramaSettingsYamlString := configMap.Data[testing.SaramaSettingsConfigKey]
 
 	// Extract (Remove) The KafkaVersion From The Sarama Config YAML
-	saramaSettingsYamlString, kafkaVersion, err := extractKafkaVersion(saramaSettingsYamlString)
+	saramaSettingsYamlString, kafkaVersion, err := client.ExtractKafkaVersion(saramaSettingsYamlString)
 	if err != nil {
 		return nil, fmt.Errorf("failed to extract KafkaVersion from Sarama Config YAML: err=%s : config=%+v", err, saramaSettingsYamlString)
 	}
 
 	// Extract (Remove) Any TLS.Config RootCAs & Set In Sarama.Config
-	saramaSettingsYamlString, certPool, err := extractRootCerts(saramaSettingsYamlString)
+	saramaSettingsYamlString, certPool, err := client.ExtractRootCerts(saramaSettingsYamlString)
 	if err != nil {
 		return nil, fmt.Errorf("failed to extract RootPEMs from Sarama Config YAML: err=%s : config=%+v", err, saramaSettingsYamlString)
 	}
